+++ src/api/staking/types.ts
